requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
cryptography==3.4.8
ebaysdk==2.2.0
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
pycparser==2.20
pyOpenSSL==20.0.1
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
